Refactor and tidy up
