--- app/controllers/steps/capital/properties_controller.rb
@@ -1,7 +1,0 @@
-module Steps
-  module Capital
-    class PropertiesController < Steps::CapitalStepController
-      include Steps::Capital::PropertyStep
-    end
-  end
-end
