+++ app/controllers/steps/capital/property_address_controller.rb
@@ -2,6 +2,20 @@ module Steps
   module Capital
     class PropertyAddressController < Steps::CapitalStepController
       include Steps::Capital::PropertyStep
+
+      private
+
+      def advance_as
+        :property_address
+      end
+
+      def form_name
+        PropertyAddressForm
+      end
+
+      def additional_permitted_params
+        [address: [:address_line_one, :address_line_two, :city, :country, :postcode]]
+      end
     end
   end
 end
