+++ app/services/decisions/capital_decision_tree.rb
@@ -15,7 +15,7 @@ module Decisions
         after_property_type(form_object.property)
       when :residential_property
         after_properties
-      when :residential_property_address
+      when :property_address
         # TODO: Route to property owner page once built
         edit(:saving_type)
       when :premium_bonds
@@ -48,7 +48,7 @@ module Decisions
     end
 
     def after_properties
-      return edit(:residential_property_address) if form_object.is_home_address.nil? || form_object.is_home_address.no?
+      return edit(:property_address) if form_object.is_home_address.nil? || form_object.is_home_address.no?
 
       # TODO: Route to appropriate property page loop once built
       edit(:saving_type) # Placeholder to join up flow
